+++ src/rendersystems/DirectX/9/DirectX9_VertexShader.cpp
@@ -41,7 +41,7 @@ namespace april
 	bool DirectX9_VertexShader::loadFile(chstr filename)
 	{
 		unsigned char* data = NULL;
-		long size = 0;
+		int32_t size = 0;
 		if (!this->_loadFileData(filename, &data, &size))
 		{
 			hlog::error(april::logTag, "Shader file not found: " + filename);
@@ -59,7 +59,7 @@ namespace april
 	bool DirectX9_VertexShader::loadResource(chstr filename)
 	{
 		unsigned char* data = NULL;
-		long size = 0;
+		int32_t size = 0;
 		if (!this->_loadResourceData(filename, &data, &size))
 		{
 			hlog::error(april::logTag, "Shader file not found: " + filename);
